+++ javaee/rt/src/java/org/aion/avm/RuntimeMethodFeeSchedule.java
@@ -284,21 +284,21 @@ public class RuntimeMethodFeeSchedule {
     public static final int BlockchainRuntime_avm_require = RT_METHOD_FEE_LEVEL_1;
     public static final int BlockchainRuntime_avm_println = RT_METHOD_FEE_LEVEL_1; // totalCost - 696; // byte code cost - 122; invoked methods cost - 574;
 
-    public static final int BlockchainRuntime_avm_hash_base = RT_METHOD_FEE_LEVEL_1;
-    public static final int BlockchainRuntime_avm_hash_per_bytes = 50;
-    public static final int BlockchainRuntime_avm_verifySignature = RT_METHOD_FEE_LEVEL_6;
-    public static final int BlockchainRuntime_avm_verifySignature_per_bytes = 50;
-    public static final int BlockchainRuntime_avm_recoverKey = RT_METHOD_FEE_LEVEL_6;
-    public static final int BlockchainRuntime_avm_recoverKey_per_bytes = 50;
-    public static final int BlockchainRuntime_avm_aggregate = RT_METHOD_FEE_LEVEL_6;
-    public static final int BlockchainRuntime_avm_aggregate_per_bytes = 20;
+    public static final long BlockchainRuntime_avm_hash_base = RT_METHOD_FEE_LEVEL_1;
+    public static final long BlockchainRuntime_avm_hash_per_bytes = 50;
+    public static final long BlockchainRuntime_avm_verifySignature = RT_METHOD_FEE_LEVEL_6;
+    public static final long BlockchainRuntime_avm_verifySignature_per_bytes = 50;
+    public static final long BlockchainRuntime_avm_recoverKey = RT_METHOD_FEE_LEVEL_6;
+    public static final long BlockchainRuntime_avm_recoverKey_per_bytes = 50;
+    public static final long BlockchainRuntime_avm_aggregate = RT_METHOD_FEE_LEVEL_6;
+    public static final long BlockchainRuntime_avm_aggregate_per_bytes = 20;
 
-    public static final int BlockchainRuntime_avm_ecAdd = RT_METHOD_FEE_LEVEL_6;
-    public static final int BlockchainRuntime_avm_ecAdd_per_points = 1000;
-    public static final int BlockchainRuntime_avm_ecScalarMul_g1 = 20000;
-    public static final int BlockchainRuntime_avm_ecScalarMul_g2 = 95000;
-    public static final int BlockchainRuntime_avm_ecPairingCheck = 200000;
-    public static final int BlockchainRuntime_avm_ecPairingCheck_per_pairs = 40000;
+    public static final long BlockchainRuntime_avm_ecAdd = RT_METHOD_FEE_LEVEL_6;
+    public static final long BlockchainRuntime_avm_ecAdd_per_points = 1000;
+    public static final long BlockchainRuntime_avm_ecScalarMul_g1 = 20000;
+    public static final long BlockchainRuntime_avm_ecScalarMul_g2 = 95000;
+    public static final long BlockchainRuntime_avm_ecPairingCheck = 200000;
+    public static final long BlockchainRuntime_avm_ecPairingCheck_per_pairs = 40000;
 
     public static final int BlockchainRuntime_avm_getAddressFromKey = RT_METHOD_FEE_LEVEL_5;
     public static final int BlockchainRuntime_avm_getFeeSharingProportion = RT_METHOD_FEE_LEVEL_1;
@@ -712,15 +712,15 @@ public class RuntimeMethodFeeSchedule {
     public static final int UnmodifiableMapEntry_equals = RT_METHOD_FEE_LEVEL_1;
     public static final int UnmodifiableMapEntry_hashCode = RT_METHOD_FEE_LEVEL_1;
 
-    public static final int ObjectWriter_writePricePerByte = 2;
-    public static final int ObjectWriter_customMethodBase = RT_METHOD_FEE_LEVEL_1;
-    public static final int ObjectWriter_beginBase = RT_METHOD_FEE_LEVEL_1;
-    public static final int ObjectWriter_endBase = RT_METHOD_FEE_LEVEL_1;
+    public static final long ObjectWriter_writePricePerByte = 2;
+    public static final long ObjectWriter_customMethodBase = RT_METHOD_FEE_LEVEL_1;
+    public static final long ObjectWriter_beginBase = RT_METHOD_FEE_LEVEL_1;
+    public static final long ObjectWriter_endBase = RT_METHOD_FEE_LEVEL_1;
 
-    public static final int ObjectReader_readPricePerByte = 2;
-    public static final int ObjectReader_skipPricePerByte = 1;
-    public static final int ObjectReader_customMethodBase = RT_METHOD_FEE_LEVEL_1;
-    public static final int ObjectReader_beginBase = RT_METHOD_FEE_LEVEL_1;
-    public static final int ObjectReader_endBase = RT_METHOD_FEE_LEVEL_1;
-    public static final int ObjectReader_hasNext = RT_METHOD_FEE_LEVEL_1;
+    public static final long ObjectReader_readPricePerByte = 2;
+    public static final long ObjectReader_skipPricePerByte = 1;
+    public static final long ObjectReader_customMethodBase = RT_METHOD_FEE_LEVEL_1;
+    public static final long ObjectReader_beginBase = RT_METHOD_FEE_LEVEL_1;
+    public static final long ObjectReader_endBase = RT_METHOD_FEE_LEVEL_1;
+    public static final long ObjectReader_hasNext = RT_METHOD_FEE_LEVEL_1;
 }
